+++ packages/ui/src/components/Tooltip/Tooltip.tsx
@@ -28,6 +28,11 @@ const VortexTooltip: React.FC<TooltipProps> = ({
               "0px 4px 6px -1px rgba(0, 0, 0, 0.1), 0px 2px 4px -1px rgba(0, 0, 0, 0.06)",
           },
         },
+        arrow: {
+          sx: {
+            color: bgColor,
+          },
+        },
       }}
       {...props}
     >