+++ packages/ui/src/theme/typography.ts
@@ -2,7 +2,7 @@ import { TypographyOptions } from '@mui/material/styles/createTypography';
 
 export const typography: TypographyOptions = {
   fontFamily: [
-    'Inter',
+    'Rubik',
     '-apple-system',
     'BlinkMacSystemFont',
     '"Segoe UI"',